backend/requirements.txt
alembic>=1.18,<2.0
anthropic>=0.67,<1.0
camel-ai==0.2.78
camel-oasis==0.2.5
fastapi>=0.115,<1.0
httpx>=0.28,<1.0
pydantic-settings>=2.8,<3.0
pypdf>=6.0,<7.0
pytest>=8.2,<9.0
pytest-asyncio>=0.23,<1.0
python-multipart>=0.0.20,<1.0
sqlalchemy>=2.0,<3.0
uvicorn[standard]>=0.34,<1.0
